requirements.txt
Django==1.11.8
gunicorn==19.7.1
whitenoise==3.3.1
ibmcloudenv
livereload
watson-developer-cloud~=2.8.0
Pillow==6.0.0
